17 implement express validator for userroute

+++ controllers/user.controller.js
@@ -1,5 +1,6 @@
 import User from "../mongodb/models/user.js";
 import bcrypt from "bcrypt";
+import mongoose from "mongoose";
 
 const getAllUsers = async (req, res) => {
   try {
@@ -12,18 +13,23 @@ const getAllUsers = async (req, res) => {
 
 const getUserById = async (req, res) => {
   try {
-    const user = await User.findById(req.params.id);
+    const { id } = req.params;
+    if (!mongoose.isValidObjectId(id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
+    const user = await User.findById(id);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
     res.status(200).json(user);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 
 const createUser = async (req, res) => {
   try {
-    if (!req.body.password || !req.body.email) {
-      return res.status(400).json({ message: "Missing password or email" });
-    }
     const hashedPassword = await bcrypt.hash(req.body.password, 10);
     const user = await User.create({ ...req.body, password: hashedPassword });
     const { password: _, ...userWithoutPassword } = user._doc;
@@ -39,21 +45,48 @@ const createUser = async (req, res) => {
 
 const updateUser = async (req, res) => {
   try {
-    const user = await User.findByIdAndUpdate(req.params.id, req.body, {
+    const { id } = req.params;
+    if (!mongoose.isValidObjectId(id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
+
+    const paramUser = await User.findById(id);
+    if (!paramUser) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    const userSameEmail = await User.findOne({ email: req.body.email });
+    if (userSameEmail && paramUser.id !== userSameEmail.id) {
+      return res.status(409).json({
+        message: "Another user is using the provided email address",
+      });
+    }
+    const user = await User.findByIdAndUpdate(id, req.body, {
       new: true,
+      runValidators: true,
     });
-    res.status(200).json(user);
+    const { password: _, ...userWithoutPassword } = user._doc;
+    res.status(200).json(userWithoutPassword);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
 
 const deleteUser = async (req, res) => {
   try {
-    const user = await User.findByIdAndDelete(req.params.id);
-    res.status(200).json(user);
+    const { id } = req.params;
+    if (!mongoose.isValidObjectId(id)) {
+      return res.status(400).json({ message: "Invalid user ID" });
+    }
+    const user = await User.findById(id);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    const userDeleted = await User.findByIdAndDelete(req.params.id);
+
+    res.status(200).json(userDeleted);
   } catch (error) {
-    res.status(404).json({ message: error.message });
+    res.status(500).json({ message: error.message });
   }
 };
